# services/base/docs/swagger.yaml
definitions:
...
  exception.ExceptionError:
    properties:
      error_subcode:
        type: integer
      message:
        type: string
      type:
        $ref: '#/definitions/exception.errorType'
    type: object
  exception.ExceptionErrors:
    properties:
      errItems:
        items:
          $ref: '#/definitions/exception.ExceptionError'
        type: array
      globalMessage:
        type: string
      httpStatusCode:
        type: integer
    type: object
  exception.errorType:
    enum:
    - UnknownError
    - BadRequest
    - InvalidUUID
    - NotFound
    - Unauthorized
    - ValidationError
    - JobError
    - ExternalServiceError
    - DatasourceError
    type: string
    x-enum-varnames:
    - ERROR_TYPE_UNKNOWN_ERROR
    - ERROR_TYPE_BAD_REQUEST
    - ERROR_TYPE_INVALID_UUID
    - ERROR_TYPE_NOT_FOUND
    - ERROR_TYPE_UNAUTHORIZED
    - ERROR_TYPE_VALIDATION_ERROR
    - ERROR_TYPE_JOB_ERROR
    - ERROR_TYPE_EXTERNAL_SERVICE_ERROR
    - ERROR_TYPE_DATASOURCE_ERROR
...
  response.CommonResponse:
    properties:
      data: {}
      errors:
        $ref: '#/definitions/exception.ExceptionErrors'
      request_id:
        type: string
      response_code:
        type: integer
      response_message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/crm/socials/instagram:
    get:
      description: Redirects the user to the Facebook OAuth consent screen to connect
        their Instagram Business account
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to Facebook OAuth
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initiate Instagram / Facebook OAuth login
      tags:
      - CRM Socials
...
  /api/v1/forms/field/types:
    get:
      description: Returns all field types with their configs and validation rules
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: List all field types
      tags:
      - Forms
  /api/v1/forms/field/types/{name}:
    get:
      description: Returns a single field type along with its configs and validation
        rules
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Field type name (URL-encoded)
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Get a field type by name
      tags:
      - Forms
...
swagger: "2.0"